day1/requirements.txt
flake8

numpy

pillow

matplotlib
